dataflow/timestamp: reallow duplicate times in add_binding
Touches #10742

This commit lets us temporarily allow adding bindings for times that already
have timestamp bindings. This occurs when a timestamp history is compacted beyond
the frontier of timestamps that have been finalized. Concretely:

1. We have a binding at t1 and a proposed binding at t2.
2. We compact up to t2. Now the binding previously at t1 is forwarded to t2.
3. At some later time, we attempt to finalize the proposed binding at t2 and
   when we do so, we observe that a binding at t2 already exists.

There's a number of potential fixes here:

1. Always flush pending bindings before doing any compaction
2. Compact up to pending_time - 1

are the two main candidates. It's not clear at the moment which is best or
if there's a third candidate that we haven't considered.