expr: rename nullary functions to unmaterializable functions
Presently all nullary functions are unmaterializable, but one day that
may not be true. Rename NullaryFunc to UnmaterializableFunc accordingly,
and similarly for associated types and methods.

The documentation is also adjusted to use the term "unmaterializable"
rather than "nonpure", per the results of a Slack discussion.
(@umanwizard felt that the term "pure" should not apply to many of the
system information functions, because their output implicitly depends on
the state of various system catalog tables.)

This completes the followup requested by @aalexandrov in the review of
PR #10449.