pgrepr: add typmod handling for date/time types
The date/time types in PostgreSQL all support precision constraints, so
our pgrepr package should as well. This commit also restructures
`pgrepr::Type::from_oid_and_typmod` so that supplying an OID to a type
that does not support it is a hard error.

To preserve the existing behavior of `CREATE VIEWS FROM SOURCE`, we
explicitly ignore the precision modifier on date/time types.