storage: fold timestamp binding commands into CreateSources
In the current state of the code we can end up into an invalid state in
the storage controller were we have cleaned up all the state related to
the source due to its frontier advancing to the empty frontier but still
have its TimestampBindingBox lying around.

We currently assert that this never happens and can lead to crashes. I
have observed it once here:

https://buildkite.com/materialize/nightlies/builds/707#5b10bc37-be92-4c00-95d2-227296e3ecdb

This patch removes the independent `{Add,Drop}SourceTimestamping`
commands from the storage controller and handles setting up and tearing
down the associated state in exactly two places. Source state is set-up
during CreateSources and and is teared down when the frontier is
advanced to the empty frontier.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>