materialized: present entire TLS certificate chain when possible
When given a TLS certificate file that contains multiple certificates
(i.e., an intermediate certificate chain), slurp them all up so that the
entire chain can be presented to clients who connect to the server.

We noticed this because `psql` was unable to connect to Materialize
Cloud deployments whose server certificates were provided by Let's
Encrypt. Let's Encrypt certificates are signed by an intermediary "R3",
whose certificates are not present in the default CA bundle. The
intermediary's certificate *is* present in the certificate returned by
Let's Encrypt, however, so we just need to teach `materialized` to
forward it along to clients.