coord: allow SELECTing from unmaterialized sources
This commit allows `SELECT` queries to depend, directly or indirectly,
on unmaterialized sources. This was previously disallowed because we
didn't track the `since` frontier of unmaterialized sources; now we do.

Users will no longer see errors like

    unable to automatically determine a query timestamp

which is a huge upgrade in the user experience, in my opinion. The
downside is that users are not careful, a one-off query may result in
e.g. re-reading an entire Kafka source from the upstream broker. If it
becomes a problem, we can consider introducing a "safe mode" (maybe via
a session variable) that disallows queries that would re-instantiate a
source.

This new behavior facilitates making tables unmaterialized by default
(#10883). That change goes down more easily if you can `SELECT` from an
unmaterialized table.