coord: introduce type to bundle storage and compute IDs
Storage and compute IDs are commonly bundled together as anonymous
tuples in parameters and return values. It is easy to accidentally
transpose the order of `storage_ids` and `compute_ids`, especially
considering our codebase flip flops on whether storage comes before
compute or vice-versa.

This commit introduces a struct to bundle these types of IDs together.
This makes it impossible to get the order wrong. It is also a good place
to add convenience methods for operating on the bundle as a whole.

Fix #11040.