coord: give logging indexes a unique ID per cluster
The catalog intends to constrain each index to a single cluster. That
means that the indexes on the logging sources cannot be hardcoded into
the binary, but must be dynamically generated for each cluster.

This commit starts down the path by dynamically generating IDs for
logging indexes at bootstrap from a new section of reserved system ID
space.