sql: Support \d <object> meta command (#10980)
This commit adds the required catalog tables, columns, and functions
required to support the \d <object> meta command. The majority of these
are not applicable to Materialize and filled with sane defaults. The
rest are implemented in a way that makes sense with Materialize.

The output of \d <object> differs slightly from Postgres because
we include primary key indexes while PostgreSQL doesn't.

Closes #9595