persist: simplify compaction test
The previous iteration of this test was too specifically coupled to the compaction
implementation in the persist crate, and made unreasonable assumptions about exactly
what happened and when.

This test is more simple, and has only two steps:

1. Write some data, and observe that the number of stored bytes increased.
2. Retract that data, and observe that the number of stored bytes decreased.

It's important to note that this happens to work because all of step 1 happened
in one batch and all of step 2 happened in one batch. If we instead, break step
1 into two insert statements, we won't observe the data reduction we expect
because now the two inserts are compacted together into one batch, which will
only be compacted together with the result of another two batches themselves
compacted together.