coord: extract index oracle type
Extract a new `ComputeInstanceIndexOracle` type which is the
authoritative reference on which indexes are available for use on a
particular compute instance. It does this by mashing up information from
the catalog and the compute controller.

This type is used to determine what indexes are available during
timestamp determination *and* dataflow building, so it fixes #8318 at
long last.

By request, this PR also fixes #10955.

With this PR in place, #11029 is unblocked. It should now be
relatively straightforward to wire up `CREATE CLUSTER`.