coord,sql: rename "default database" to "active database"
The latter better represents that the database is the actively-selected
database. (It is occasionally *used as a default* when none is
specified, but it is a meaningfully different concept from the default
database being the database named "materialize".)

This commit also makes the analogous rename from "user" to "active
user".