sql,coord: allow specifying the cluster of an index or sink
Add an `IN CLUSTER` clause to `CREATE INDEX` and `CREATE SINK`, as in:

    CREATE INDEX foo IN CLUSTER bar ON table (col)

This allows creating indexes in clusters besides the active cluster.
Crucially, it is also what allows the catalog to record in which cluster
an index or sink belongs.

As with table names, cluster names are resolved to identifiers before
being stored in the catalog. For example, the above DDL statement gets
rewritten to:

    CREATE INDEX foo IN CLUSTER [1] ON table (col)

The above ugliness is visible in `SHOW CREATE INDEX` for now;
unresolving the ID `[1]` back into a name like `bar` is left to a future
commit.

Co-authored-by: Sean Loiselle <sean@materialize.io>