dataflow-types: partition StorageCommand::Insert by hash of data
Previously the inserts were partitioned by their position in the
command. This impeded compaction, because it was likely that the
retraction of a piece of data would land on a different worker than the
the original presentation of that data landed on.

This glitch was responsible for the ever-increasing time to `UPDATE` a
nonmaterialized table.

Fix #11071.