dataflow-types: avoid unsafe stream generation
As spotted by @petrosagg in #11223, the `SelectStream` incorrectly
converts `Client`s to `Stream`s by constructing new futures on every
call to `poll_next`. The code happens to work with the current stack of
futures/streams, but might break at any point in the future.

The solution here involves many allocations on each call to `recv()`
but avoids the defect. If these allocations become a performance
problem, we can look into a more invasive refactor like #11314.