coordtest: remove
Take @mjibson up on his offer to wholesale remove coordtest. Several
different PRs (#11314, #11318) would have required hairy refactors to
coordtest and so opted to remove it. Both of those PRs are stalled out,
but they'll be landing soon in modified form, so just get the coordtest
removal out of the way.

To be clear, coordtest is a good idea, and one we should rebuild as soon
as we've got the improved APIs in place. It should be much easier to
build coordtest in this new world, too, with clear API boundaries
between the ADAPTER, STORAGE, and COMPUTE layers.