feature-benchmark: Tone down the StartupLoaded scenario
Due to #11332 there is a limit to the number of persisted objects
one can have or Mz will fail to restart.

As we run with the same SCALE for both the non-persisted and
persisted invocations of the benchmark, we have no option at this time
but to down down both.

Also, make sure that we only declare the restart as complete after
all the objects we created pre-restart have become queryable.