orchestrator: add initial managed cluster implementation
This commit teaches `materialized` to optionally manage its own compute
clusters. This functionality is provided by two new crates:

  * The `orchestrator` crate provides an abstraction for managing
    services, which are a set of one or more processes running the same
    image.

  * The `orchestrator-kubernetes` crate provides a concrete
    implementation of the `Orchestrator` trait that is backed by
    Kubernetes. A service in the `KubernetesOrchestrator` corresponds to
    a StatefulSet in Kubernetes.

In the future, we will likely provide additional orchestrator
implementations that manage processes on the local machine or images in
a Docker daemon to ease local development. This commit adds only the
Kubernetes-backed implementation since that is what we plan to use in
production.

Local development of the orchestrator is facilitated by the addition of
a family of scripts which build and deploy `materialized` to Minikube.
The scripts are:

  * `bin/cluster-dev`, which deploys `materialized` into Minikube,
    essentially simulating the environment controller locally
  * `bin/cluster-psql`, which launches a psql shell connected to the
    `materialized` instance running in Minikube
  * `bin/cluster-reset`, which blows away all Minikube state

This is prototype quality code. There are many missing retry loops, and
the `materialized` that gets deployed to Kubernetes crash loops several
times before it becomes healthy. And there are no tests. But none of
this code is exercised in the default configuration, and using it is
gated by `--experimental`.