sql,coord: add prototype of ALTER CLUSTER
ALTER CLUSTER is a possibly short-lived command to add and remove
replicas of clusters. Specifically, the syntax of CREATE CLUSTER ...
REMOTE is changed to

    CREATE CLUSTER <name> [REMOTE <name> (<host>, <host> ...) ...]

where multiple REMOTE clauses may now be specified to name multiple
replicas. A nearly identical ALTER CLUSTER command

    ALTER CLUSTER <name> [REMOTE <name> (<host>, <host> ...) ...]

can we used to wholesale replace the remote replica set of a cluster.
Replicas are identified by the user provided name. After alteration, any
new replicas are added and any unnamed replicas are removed. Changing
the host specification of an existing replica is not permitted.

Forthcoming work will make the concept of a replica more explicitly part
of the system (e.g., by allowing virtual and managed replicas). But for
now this unblocks the active replication work.