partitioned client: Partition state by instance
Currently, the state of a partition client is shared across all instances.
This can cause cross-talk between instances, especially when dropping an
instance causes all state to be reset. We avoid the problem by maintaining
state per compute instance, which permits us to maintain the state with a
finer granularity.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>