antithesis: Run Mz in the Antithesis environment with --workers 2
Previously, we ran it with --workers 1 because the docker-compose.yml
file was created eons ago, which did not provide any concurrency on the
timely/dataflow level.

Unfortunately, running with --workers 4 causes too many queries to never
complete, so --workers 2 is currently the only viable settings to get
at least some concurrency going.