ore: compensate for changed retry default max duration
PR #11425 somewhat inadvertently changed `Retry::default()` so that it
produces a retry operation that retries forever, rather than a retry
operation that has a maximum duration of 30s.

This is actually a better default, because the default is the *least*
constrained retry operation, and each caller can choose whether to add a
max try constraint, a max duration constraint, or both. Previously
callers that wanted to retry forever had to know to call
`max_tries(usize::MAX)` to remove the max duration constraint.

So, this commit keeps the new default, but updates all callers to
account for the new default. Callers that want a 30s max duration need
to install it explicitly; callers that want an infinite retry loop no
longer need to call `max_tries(usize::MAX)`.

h/t @nharring-adjacent for spotting this.