ore/retry: improve semantics of canceling retry errors
Previously if a canceling retry operation timed out the underlying
function, the error returned was always "deadline exceeded." This is not
very informative. In most cases, we can do better: if the underlying
function has previously failed with an error message, that error message
is likely quite a bit more useful than "deadline exceeded", and we
should return it instead.